+++ tests/controllers/test_joining.py
@@ -85,7 +85,7 @@ class TestJoin(unittest.TestCase):
 
     def test_has_attributes(self):
         attributes = [
-            "evefile",
+            "file",
         ]
         for attribute in attributes:
             with self.subTest(attribute=attribute):
@@ -94,17 +94,17 @@ class TestJoin(unittest.TestCase):
     def test_initialise_with_evefile_sets_evefile(self):
         evefile = self.evefile
         join = joining.Join(file=evefile)
-        self.assertEqual(evefile, join.evefile)
+        self.assertEqual(evefile, join.file)
 
     def test_join_without_evefile_raises(self):
-        self.join.evefile = None
+        self.join.file = None
         with self.assertRaisesRegex(
             ValueError, "Need an evefile to join data."
         ):
             self.join.join()
 
     def test_join_without_data_raises(self):
-        self.join.evefile = self.evefile
+        self.join.file = self.evefile
         with self.assertRaisesRegex(ValueError, "Need data to join data."):
             self.join.join()
 
@@ -114,8 +114,8 @@ class TestJoin(unittest.TestCase):
                 return data
 
         self.join = MyJoin()
-        self.join.evefile = self.evefile
-        self.join.evefile.data = {
+        self.join.file = self.evefile
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -132,8 +132,8 @@ class TestJoin(unittest.TestCase):
                 return data
 
         self.join = MyJoin()
-        self.join.evefile = self.evefile
-        self.join.evefile.data = {
+        self.join.file = self.evefile
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -141,50 +141,8 @@ class TestJoin(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
         }
-        result = self.join.join(data=list(self.join.evefile.data.values()))
+        result = self.join.join(data=list(self.join.file.data.values()))
         self.assertTrue(result)
-        for item in result:
-            self.assertIsInstance(item, evefile.entities.data.MeasureData)
-
-    def test_join_with_ids_converts_to_datasets(self):
-        class MyJoin(joining.Join):
-            def _join(self, data=None):
-                return data
-
-        self.join = MyJoin()
-        self.join.evefile = MockEveFile()
-        self.join.evefile.data = {
-            "SimChan:01": MockChannel(
-                data=np.random.random(5), positions=np.linspace(0, 4, 5)
-            ),
-            "SimMot:01": MockAxis(
-                data=np.random.random(7), positions=np.linspace(0, 6, 7)
-            ),
-        }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
-        for item in result:
-            self.assertIsInstance(item, evefile.entities.data.MeasureData)
-
-    def test_join_with_names_converts_to_datasets(self):
-        class MyJoin(joining.Join):
-            def _join(self, data=None):
-                return data
-
-        self.join = MyJoin()
-        self.join.evefile = MockEveFile()
-        self.join.evefile.data = {
-            "SimChan:01": MockChannel(
-                data=np.random.random(5),
-                positions=np.linspace(0, 4, 5),
-                name="bar",
-            ),
-            "SimMot:01": MockAxis(
-                data=np.random.random(7),
-                positions=np.linspace(0, 6, 7),
-                name="baz",
-            ),
-        }
-        result = self.join.join(data=["bar", "baz"])
         for item in result:
             self.assertIsInstance(item, evefile.entities.data.MeasureData)
 
@@ -192,31 +150,41 @@ class TestJoin(unittest.TestCase):
 class TestChannelPositions(unittest.TestCase):
     def setUp(self):
         self.join = joining.ChannelPositions()
-        self.join.evefile = MockEveFile()
+        self.join.file = MockEveFile()
 
     def test_instantiate_class(self):
         pass
 
     def test_join_returns_list_of_measuredata(self):
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         for item in result:
             self.assertIsInstance(item, evefile.entities.data.MeasureData)
 
     def test_join_returns_copy_of_measuredata(self):
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
-        self.assertIsNot(result[0], self.join.evefile.data["SimChan:01"])
-        self.assertIsNot(result[1], self.join.evefile.data["SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
+        self.assertIsNot(result[0], self.join.file.data["SimChan:01"])
+        self.assertIsNot(result[1], self.join.file.data["SimMot:01"])
 
     def test_join_returns_data_in_input_order(self):
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(
+            data=[
+                self.join.file.data["SimMot:01"],
+                self.join.file.data["SimChan:01"],
+            ]
+        )
         self.assertIsInstance(result[0], evefile.entities.data.AxisData)
         self.assertIsInstance(result[1], evefile.entities.data.ChannelData)
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(
+            data=[
+                self.join.file.data["SimChan:01"],
+                self.join.file.data["SimMot:01"],
+            ]
+        )
         self.assertIsInstance(result[0], evefile.entities.data.ChannelData)
         self.assertIsInstance(result[1], evefile.entities.data.AxisData)
 
     def test_join_returns_only_values_for_channel_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
@@ -224,14 +192,14 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         for item in result:
             self.assertEqual(
-                len(self.join.evefile.data["SimChan:01"].data), len(item.data)
+                len(self.join.file.data["SimChan:01"].data), len(item.data)
             )
 
     def test_join_returns_values_for_all_channel_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -239,14 +207,14 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         for item in result:
             self.assertEqual(
-                len(self.join.evefile.data["SimChan:01"].data), len(item.data)
+                len(self.join.file.data["SimChan:01"].data), len(item.data)
             )
 
     def test_join_returns_only_values_for_channel_positions_with_gaps(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(4), positions=np.asarray([0, 2, 3, 4])
             ),
@@ -254,11 +222,11 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
         }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
 
     def test_join_fills_axes_values(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -266,16 +234,16 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(4), positions=np.linspace(0, 3, 4)
             ),
         }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
         np.testing.assert_array_equal(
-            self.join.evefile.data["SimMot:01"].data,
+            self.join.file.data["SimMot:01"].data,
             result[1].data[:-1],
         )
         self.assertEqual(result[1].data[-2], result[1].data[-1])
 
     def test_join_masks_axes_values_with_gap_at_beginning(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -283,17 +251,17 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(3), positions=np.linspace(2, 4, 3)
             ),
         }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
         np.testing.assert_array_equal(
-            self.join.evefile.data["SimMot:01"].data,
+            self.join.file.data["SimMot:01"].data,
             result[1].data[2:],
         )
         self.assertIsInstance(result[1].data, ma.masked_array)
         self.assertTrue(result[1].data.mask[0])
 
     def test_join_fills_axes_values_with_gaps(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -301,13 +269,13 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(4), positions=np.asarray([0, 2, 4])
             ),
         }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
         self.assertEqual(result[1].data[0], result[1].data[1])
         self.assertEqual(result[1].data[2], result[1].data[3])
 
     def test_join_fills_axes_values_with_snapshots(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.arange(2, 7)
             ),
@@ -315,7 +283,7 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(4), positions=np.asarray([3, 4, 5, 6])
             ),
         }
-        self.join.evefile.snapshots = {
+        self.join.file.snapshots = {
             "SimChan:01": MockChannel(
                 data=np.random.random(2), positions=np.asarray([1, 7])
             ),
@@ -323,16 +291,16 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(2), positions=np.asarray([1, 7])
             ),
         }
-        self.join.evefile.set_ids()
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        self.join.file.set_ids()
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
         self.assertEqual(
-            self.join.evefile.snapshots["SimMot:01"].data[0],
+            self.join.file.snapshots["SimMot:01"].data[0],
             result[1].data[0],
         )
 
     def test_join_returns_values_for_union_of_all_channel_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -343,19 +311,17 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(5), positions=np.linspace(2, 6, 5)
             ),
         }
-        result = self.join.join(
-            data=["SimMot:01", "SimChan:01", "SimChan:02"]
-        )
+        result = self.join.join(data=self.join.file.data.values())
         np.testing.assert_array_equal(
             result[0].position_counts,
             np.union1d(
-                self.join.evefile.data["SimChan:01"].position_counts,
-                self.join.evefile.data["SimChan:02"].position_counts,
+                self.join.file.data["SimChan:01"].position_counts,
+                self.join.file.data["SimChan:02"].position_counts,
             ),
         )
 
     def test_join_fills_and_masks_channel_values_with_gaps(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -366,9 +332,7 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(5), positions=np.linspace(2, 6, 5)
             ),
         }
-        result = self.join.join(
-            data=["SimMot:01", "SimChan:01", "SimChan:02"]
-        )
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
         self.assertEqual(len(result[0].data), len(result[2].data))
         self.assertIsInstance(result[1].data, ma.masked_array)
@@ -378,13 +342,13 @@ class TestChannelPositions(unittest.TestCase):
         np.testing.assert_array_equal(
             result[1].position_counts,
             np.union1d(
-                self.join.evefile.data["SimChan:01"].position_counts,
-                self.join.evefile.data["SimChan:02"].position_counts,
+                self.join.file.data["SimChan:01"].position_counts,
+                self.join.file.data["SimChan:02"].position_counts,
             ),
         )
 
     def test_join_with_monitors(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -392,9 +356,9 @@ class TestChannelPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
         }
-        data = list(self.join.evefile.data.values())
-        mapper = timestamp_mapping.Mapper(self.join.evefile)
-        device_data = mapper.map(list(self.join.evefile.monitors.keys())[0])
+        data = list(self.join.file.data.values())
+        mapper = timestamp_mapping.Mapper(self.join.file)
+        device_data = mapper.map(list(self.join.file.monitors.keys())[0])
         data.append(device_data)
         result = self.join.join(data=data)
         self.assertTrue(result)
@@ -409,13 +373,13 @@ class TestChannelPositions(unittest.TestCase):
 class TestAxisPositions(unittest.TestCase):
     def setUp(self):
         self.join = joining.AxisPositions()
-        self.join.evefile = MockEveFile()
+        self.join.file = MockEveFile()
 
     def test_instantiate_class(self):
         pass
 
     def test_join_returns_only_values_for_axis_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -423,14 +387,14 @@ class TestAxisPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         for item in result:
             self.assertEqual(
-                len(self.join.evefile.data["SimMot:01"].data), len(item.data)
+                len(self.join.file.data["SimMot:01"].data), len(item.data)
             )
 
     def test_join_returns_values_for_all_axis_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
@@ -438,14 +402,14 @@ class TestAxisPositions(unittest.TestCase):
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         for item in result:
             self.assertEqual(
-                len(self.join.evefile.data["SimMot:01"].data), len(item.data)
+                len(self.join.file.data["SimMot:01"].data), len(item.data)
             )
 
     def test_join_returns_only_values_for_axis_positions_with_gaps(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimChan:01": MockChannel(
                 data=np.random.random(7), positions=np.linspace(0, 6, 7)
             ),
@@ -453,20 +417,20 @@ class TestAxisPositions(unittest.TestCase):
                 data=np.random.random(4), positions=np.asarray([0, 2, 3, 4])
             ),
         }
-        result = self.join.join(data=["SimChan:01", "SimMot:01"])
+        result = self.join.join(data=self.join.file.data.values())
         self.assertEqual(len(result[0].data), len(result[1].data))
 
 
 class TestAxisAndChannelPositions(unittest.TestCase):
     def setUp(self):
         self.join = joining.AxisAndChannelPositions()
-        self.join.evefile = MockEveFile()
+        self.join.file = MockEveFile()
 
     def test_instantiate_class(self):
         pass
 
     def test_join_returns_only_values_for_intersection_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -474,33 +438,33 @@ class TestAxisAndChannelPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(2, 8, 7)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         positions = np.intersect1d(
-            self.join.evefile.data["SimChan:01"].position_counts,
-            self.join.evefile.data["SimMot:01"].position_counts,
+            self.join.file.data["SimChan:01"].position_counts,
+            self.join.file.data["SimMot:01"].position_counts,
         ).astype(np.int64)
         for item in result:
             self.assertEqual(len(positions), len(item.data))
         np.testing.assert_array_equal(
             result[0].data,
-            self.join.evefile.data["SimMot:01"].data[2:6],
+            self.join.file.data["SimMot:01"].data[2:6],
         )
         np.testing.assert_array_equal(
             result[1].data,
-            self.join.evefile.data["SimChan:01"].data[0:3],
+            self.join.file.data["SimChan:01"].data[0:3],
         )
 
 
 class TestAxisOrChannelPositions(unittest.TestCase):
     def setUp(self):
         self.join = joining.AxisOrChannelPositions()
-        self.join.evefile = MockEveFile()
+        self.join.file = MockEveFile()
 
     def test_instantiate_class(self):
         pass
 
     def test_join_returns_only_values_for_union_positions(self):
-        self.join.evefile.data = {
+        self.join.file.data = {
             "SimMot:01": MockAxis(
                 data=np.random.random(5), positions=np.linspace(0, 4, 5)
             ),
@@ -508,10 +472,10 @@ class TestAxisOrChannelPositions(unittest.TestCase):
                 data=np.random.random(7), positions=np.linspace(2, 8, 7)
             ),
         }
-        result = self.join.join(data=["SimMot:01", "SimChan:01"])
+        result = self.join.join(data=self.join.file.data.values())
         positions = np.union1d(
-            self.join.evefile.data["SimChan:01"].position_counts,
-            self.join.evefile.data["SimMot:01"].position_counts,
+            self.join.file.data["SimChan:01"].position_counts,
+            self.join.file.data["SimMot:01"].position_counts,
         )
         for item in result:
             self.assertEqual(len(positions), len(item.data))
@@ -541,4 +505,4 @@ class TestJoinFactory(unittest.TestCase):
     def test_get_join_with_evefile_sets_evefile(self):
         self.factory.file = "foo"
         join = self.factory.get_join()
-        self.assertEqual(self.factory.file, join.evefile)
+        self.assertEqual(self.factory.file, join.file)